requirements.txt
protobuf
pyserial